Separate ifetch from PC, fetch compressed instr, refactor I$
The current pc counter was just a stub to have an input to the ifetch
module, but the ifetch is the module that knows where the next
instruction to fetch is (variable size/compressed instr, unconditional
jumps, branch pred, etc).
The program counter just tracks the next to-be-retired
instruction's address (used as a base for relative branches, AUIPC,
and for misc bookkeeping).

Due to this rewrite we now properly support fetching compressed
instructions and 16bit aligned instructions in general.
Also moving the internal details of the icache in its own module,
this makes it a little more parametrizable.