Add mtvec and support trapping to it on exceptions
Instead of the previous "just deadlock exec" exception handling.
CSRs now have a write-mask to protect the reserved mode bit in mtvec
(this is the "WARL" behavior required by the spec for some CSR fields)

Note that we internally keep track of the exception cause,
but this is not yet exposed (no mcause).
There is also no MRET yet, no mepc, etc. Just trapping to mtvec.