Teach ifetch to detect and speculatively follow some simple branches
This is the most basic form of 'branch prediction',
we just predict any backwards branch is taken.
As a result a short 3 instr loop testbench takes ~3x less cycles.

However this is a pretty naive implementation at every level,
and it adds a lot of delay to the ifetch.

Eventually, we should split the ifetch in two stages and make it
output >1 instr/cycle into a FIFO when possible, so that following
branches can skip a cycle without that delay being visible downstream