Do Zmm instructions in two cycles for a lot of Fmax
Yes, it's still a dumb builtin * operator!
But the critical path was going from the bypass output value,
(which is actually combinatorially from the LSU dcache...),
then to the DSPs, then a giant MUX for the exec_int output.

Turns out just putting a reg in between, nominally after the DSP,
cuts the path close enough to the middle.
I don't know if Vivado is clever enough to do retiming in the middle
of that * operator, but I sure hope so because implementing
a fast binary multiplication by hand actually sounds like work :)