exec_mem: Tweak loaded_data comb path for slightly less WNS, more TNS
On balance, I think I'm happier with a little less negative slack,
even if more paths are affected.
I'd rather have a design that just barely closes timing on thousands
of paths, than something where most paths have slack but one of them
pushes the slack further than the other and affects Fmax.

But I'm not really sure about any of that, so at least it'll be in
the Git history and we can always try to compare again later =)