Bump PLL sys_clk to 50MHz from 25MHz (we close timing up to 65)
It just barely closes timing at 65MHz, but really that's okay :)
It can be fun to claw back a little bit of slack here and there,
but I don't want to spend too much time optimizing and reworking the
pipeline until it runs at a clean 100MHz.

So long as it's not absurdly slow, I'd rather add more features
and if the pipeline has awful timings and is full of bubbles... yep :)

I do want to make sure we keep closing timing at 50MHz on the arty
board as a baseline, but that's all.