crit path: Two cycle fixp mults for integer division
We have a 66 * 128 bit mult, then some swizzling and a 127 bit add
all combinatorial in the same cycle.
So that was not going to clock very high =]
Split the mult into two halves done separately in the first cycle,
accumulate them and do the rest of the work in a second cycle.

This makes our divider make twice as many cycles, but that's fine.
Division by non-constants is slow, just don't do that in hot loops!