verilog-ethernet based AXI4-Lite echo device
This is a very dumb ethernet echo device. You send a frame to a
broadcast MAC address, it echoes it right back to you with the
broadcast as the _source_ address!
That amuses Wireshark, who highlights the ethernet frames yellow.

The device is connected to the bus, but it can't be used to read/write
anything yet, it just echoes everything until it gets reset.

Also, I'm only targetting the Artix7 board, but in principle the
verilog-ethernet library seems to be pretty portable, which is nice.