Fix inverted dcache update logic on writes
Pretty scary that tests didn't catch this until now!
I think it's because we have mostly read/write tests that work at
the level of the AXI interface, to test that the devices are working.
We do have a bunch of functional tests, but I guess never hit the
case of 1) priming the cache, 2) writing a value, 3) reading at the
same address.
Most of our functional tests write results at different addresses
so we can check all the results are correct, or they write and read
back, but without the cache having been primed first, which means
the write doesn't actually hit the cache, and the read goes to memory!

Conclusion: we'll need more functional tests that exercise the full
core by giving it real instructions...
This will be much easier once I write the little piece of
infrastructure to write tests in C/Rust and load the compiled
machine code in a verilog testbench harness directly.
(Possibly, I should run RISC-V compliance test suites directly)