Remove no longer useful ice40 board and spi_soc
This definitely has not been able to run on an ice40 in a LONG time!
We blew past the LUT budget a long time ago, even in 32 bit, with
a reduced ALEN and with BRAM caches instead of distributed RAM.

The little SPI-based debug module with a custom protocol is also
completely unused and has been for a while. Having to run a bunch
of wires manually to the raspi GPIO and communicate with it in python
got old pretty fast :)
Now that ethernet is starting to work, the next time I need a debug
interface, I'll have a much better replacement!