Boot payload over ethernet
Pretty limited right now since ethernet max payload size is 1500 B,
and with the xxhash dependency we're at 2.4kB ^^'

Works fine in tests if I just build a payload without xxhash and
boot that over the network. Since it's the loader that needs
to check the hash and not the payload, the hash is still checked :)

Now that hash is somewhat redundant with the Ethernet frame's
32 bit CRC/FSC and will become even more redundant when I move
from raw Ethernet frames to IPv4. But it's still good to have!
If it turns out the FPGAs are receiving corrupt frames that go through
the CRC for whatever reason, I'd rather know about it than spend
who knows how long trying to debug it...
Probably unlikely to happen, but the hash still pulls its weight
by making that failure mode definitely out of the realm of possible.