basic_cache: Remove buggy bram-based implementation
It seems like the ifetch is causing a lot of trouble, code is not
executing as it should when the icache is turned on, and starts
working again when turned off or replace with the non-bram impl.

To save myself a debugging session for legacy iCE40 code that was
always needlessly complicated, it can just be removed.

The non-manual-bram-instantation implementation has another quirk,
because it reads combinatorially from the memory. That prevents it
from becoming a hard BRAM block, but that happens to be better on
balance for the dcache (reading combinatorially in the input cycle
saves us some combinatorial path on the next output cycle to check
whether the line was valid)

For the ifetch it doesn't matter too much either way. This just
consumes for lutram/distributed RAM, but there's plenty spare.

Future improvement could be to add back a mode where it can still
synthesize a hard BRAM block, just not the crazy iCE40 implementation!