Placeholder 56bit datapath ethernet IP modules from 64bit wide versions
This only creates placeholder 56bit modules that are a copy of the
64bit wide modules, so only the name differs.
This is so we can diff the state with the next commit easily to
keep track of changes there are between the 64 and 56 bit versions.

Upstream provides 8 bit or 64 bit, while the lower level modules
are fully parametrized, so this has to be its own set of modules.