Teach ethernet module receive and echo IP packets
The ethernet MMIO interface is disconnected as of this commit,
as the device is now temporarily hardwired as an echo server.

Later, we can wire the MMIO interface again, without really breaking
the MMIO protocol, but adding new registers so that it sends IP
packets instead of raw Ethernet frames.