Use regular BRAM for icache, don't check tags early
We had this weird optimiztion for the dcache, and after removing
the previous buggy BRAM based icache, it started using that too.
That was fine, except the timings got a lot worse, ifetch doesn't
have a lot of slack at the end of a cycle, but is already waiting
on a lot of stuff at the start, so it's much better to use a
normal basic BRAM block and check the tag at the start of a cycle.