Always register exec_mem output
The sillyness with trying to save a cycle by outputting
some exec_mem results combinatorially was.. not good, it turns out.
Having a combinatorial path from the dcache all the way through
exec_mem's swizzling, to exec's bypass path for registers,
all the way into the next cycle instruction's dependency... yeah.
That was not good.