rust: Add network boot function again, but now over UDP
Instead of exchanging raw ethernet frames like we did before,
which limited us to the 1500 MTU without workaround,
we have a nice UDP stream that accumulates into the rx buffer.
The protocol is just length, checksum, and payload sent in datagrams
of maximum 1500 bytes (MTU).
The device ACKs each datagram before the full message, this allows
us to not overflow the hardware RX buffer which is actually not
that big (2K).