rust: Rewrite panic/trap handling to reliably report failures
Previously the trap handler was the panic handler, but since
we build with panic_immediate_abort that made no sense.
panic!() never called the panic handler with panic information,
instead if executed an illegal instruction and the trap handler
got called.

This made no sense because the trap handler doesn't get a panic info
arg, and the panic handler is never called anyways.
So now we have a proper trap handler and an empty panic handler.