axi4lite_sram: Avoid using the bram module
The bram module is a relic from the ice40 days,
trying to enforce exactly how the memory will be split into
ice40 physical BRAMs.
It's inflexible and just not great to use.

Let's just let the P&R tool figure it out,
and if our big memory blocks don't get mapped to BRAM,
surely we're not going to miss it easily =)