SRLW/SRLIW: Fix edge-case of sign-extending zero-shifted signed input
SRAIW/SRAW sign-extend the shifted intermediate number
to 32 bits, while the SRLIW/SRLW versions don't.

But these four are all defined to work on a 32-bit number,
and THEN always sign-extend the result to 64-bit.

So in the edge case of a logical shift by 0 of a signed input,
the intermediate 32-bit result will still have a sign bit set,
and then must be sign-extended to 64-bit.

We missed this.