Fix exec_div_log2 unique if QuestaSim warning
The fix is stupid, the logic generated will be the same.
But eliminating the if ('0) makes Questa happy.

The incorrect warning is that no branch is true.
In fact, it seems to be unhappy because the first branch
is always false, even if one of the other branches was true.